% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ginidecomp.R
\name{Ginidecomp}
\alias{Ginidecomp}
\title{Gini Decomposition Function}
\usage{
Ginidecomp(x, z, w = NULL, param = 2)
}
\arguments{
\item{param}{is the segments}
}
\description{
This Function allows you to calculate Gini coefficient decomposition.
}
\examples{
decompSGini(income, gender, param=2)
decompSGini()
}
\keyword{Decomposition}
\keyword{Gini}
\keyword{Inequality}

